Moves objects around
