+++ src/cli/definitions/parse-server.js
@@ -1,2 +1,2 @@
-const ParseServerDefinitions = require('../../Definitions').ParseServerCLIOptions;
+const ParseServerDefinitions = require('../../Options/Definitions').ParseServerOptions;
 export default ParseServerDefinitions;
